feat: 상품 목록 -> 상세 페이지 연결 (Tab + ViewPager2)
#28
