+++ app/src/main/java/com/example/bundlebundle/product/detail/ProductDetailActivity.kt
@@ -6,12 +6,13 @@ import android.util.Log
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.fragment.app.Fragment
 import com.bumptech.glide.Glide
 import com.example.bundlebundle.BottomSheetPurchaseFragment
 import com.example.bundlebundle.R
-import com.example.bundlebundle.databinding.ActivityProductDetailBinding
 import com.example.bundlebundle.retrofit.ApiClient
 import com.example.bundlebundle.retrofit.dataclass.ProductVO
+import com.example.bundlebundle.template.UpbuttonTemplateActivity
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -22,23 +23,11 @@ import kotlin.coroutines.resumeWithException
 import kotlin.coroutines.suspendCoroutine
 import kotlin.properties.Delegates
 
-class ProductDetailActivity : AppCompatActivity() {
-
-    private lateinit var openBottomSheetButton: Button
-    private lateinit var productImageView: ImageView
-    private lateinit var productNameTextView: TextView
-    private lateinit var productPriceTextView: TextView
-    private lateinit var productOriginTextView: TextView
-    private lateinit var productPackageTypeTextView: TextView
-
-    private val apiService = ApiClient.apiService
+class ProductDetailActivity : UpbuttonTemplateActivity() {
 
     private var productId by Delegates.notNull<Int>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        val binding = ActivityProductDetailBinding.inflate(layoutInflater)
-        setContentView(binding.root)
 
         productId = intent.getIntExtra("productId", -1)
         if (productId == -1) {
@@ -46,56 +35,10 @@ class ProductDetailActivity : AppCompatActivity() {
             return
         }
 
-        initializeViews(binding)
-        setupButtonListeners()
-        getProductFromApi(productId)
+        super.onCreate(savedInstanceState)
     }
 
-    private fun initializeViews(binding: ActivityProductDetailBinding) {
-        openBottomSheetButton = binding.buttonOpenBottomSheet
-        productImageView = binding.productImageView
-        productNameTextView = binding.productFullNameTextView
-        productPriceTextView = binding.productPriceTextView
-        productOriginTextView = binding.productOriginTextView
-        productPackageTypeTextView = binding.productPackageTypeTextView
-    }
-
-    private fun setupButtonListeners() {
-        openBottomSheetButton.setOnClickListener {
-            openBottomSheet()
-        }
-    }
-
-    private fun getProductFromApi(productId: Int) {
-        apiService.showProductDetail(productId).enqueue(object : Callback<ProductVO> {
-            override fun onResponse(call: Call<ProductVO>, response: Response<ProductVO>) {
-                if (response.isSuccessful) {
-                    val product = response.body()
-                    bindProductData(product)
-                } else {
-                    Log.e("ProductDetailActivity", "Failed to fetch product from API.")
-                }
-            }
-
-            override fun onFailure(call: Call<ProductVO>, t: Throwable) {
-                Log.e("ProductDetailActivity", "API call failed.", t)
-            }
-        })
-    }
-
-    private fun bindProductData(product: ProductVO?) {
-        product?.let {
-            Glide.with(this).load(it.thumbnailImg).into(productImageView)
-            productNameTextView.text = it.name
-            val priceFormatted: String = NumberFormat.getNumberInstance(Locale.getDefault()).format(it.price)
-            productPriceTextView.text = priceFormatted
-            productOriginTextView.text = it.origin
-            productPackageTypeTextView.text = it.packageType
-        }
-    }
-
-    private fun openBottomSheet() {
-        val bottomSheetFragment = BottomSheetPurchaseFragment()
-        bottomSheetFragment.show(supportFragmentManager, "bottomSheet")
+    override fun setFragment(): Fragment {
+        return ProductDetailFragment.newInstance(productId)
     }
 }
