+++ app/src/main/java/com/example/bundlebundle/product/list/ProductGridFragment.kt
@@ -39,9 +39,7 @@ class ProductGridFragment : Fragment() {
         _binding = FragmentProductGridBinding.inflate(inflater, container, false)
 
         bindWithApiResponse()
-
         binding.productGridList.layoutManager = GridLayoutManager(requireContext(), columnCount)
-
         return binding.root
     }
