+++ app/src/main/res/layout/fragment_product_detail.xml
@@ -41,9 +41,9 @@
 
                 <TextView
                     android:id="@+id/product_price_text_view"
-                    android:layout_width="100dp"
+                    android:layout_width="wrap_content"
                     android:layout_height="50dp"
-                    android:gravity="right"
+                    android:gravity="left"
                     android:paddingLeft="15dp"
                     android:paddingTop="15dp"
                     android:text="28,000"
